добавление картинок для объявлений, исправление замечаний

+++ src/main/java/ru/skypro/homework/dto/ResponseWrapperAdsDto.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class ResponseWrapperAdsDto {
-    private long count;
+    private Integer count;
     private List<AdsDto> results;
 }
